refactor(app): getLabwareRenderInfo into xforms
