+++ app/src/organisms/Devices/ProtocolRun/SetupLabware/__tests__/SetupLabwareMap.test.tsx
@@ -15,7 +15,7 @@ import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 import { getAttachedProtocolModuleMatches } from '../../../../ODD/ProtocolSetup/ProtocolSetupModulesAndDeck/utils'
 import { LabwareInfoOverlay } from '../../LabwareInfoOverlay'
-import { getLabwareRenderInfo } from '../../utils/getLabwareRenderInfo'
+import { getLabwareRenderInfo } from '/app/transformations/analysis'
 import { SetupLabwareMap } from '../SetupLabwareMap'
 
 import type {
@@ -42,7 +42,7 @@ vi.mock('@opentrons/shared-data', async importOriginal => {
 
 vi.mock('../../../../ODD/ProtocolSetup/ProtocolSetupModulesAndDeck/utils')
 vi.mock('../../LabwareInfoOverlay')
-vi.mock('../../utils/getLabwareRenderInfo')
+vi.mock('/app/transformations/analysis/getLabwareRenderInfo')
 vi.mock('../../utils/getModuleTypesThatRequireExtraAttention')
 vi.mock('../../../../RunTimeControl/hooks')
 vi.mock('../../../hooks')
