+++ app/src/organisms/ODD/ProtocolSetup/ProtocolSetupModulesAndDeck/__tests__/ModulesAndDeckMapView.test.tsx
@@ -19,7 +19,7 @@ vi.mock('/app/redux/config')
 vi.mock('../../../../Devices/hooks')
 vi.mock('/app/resources/deck_configuration/utils')
 vi.mock('../../../../Devices/ModuleInfo')
-vi.mock('../../../../Devices/ProtocolRun/utils/getLabwareRenderInfo')
+vi.mock('/app/transformations/analysis')
 
 const mockRunId = 'mockRunId'
 const PROTOCOL_ANALYSIS = {
