+++ app/src/transformations/analysis/getLabwareRenderInfo.ts
