+++ app/src/transformations/analysis/index.ts
@@ -1,3 +1,4 @@
 export * from './parseProtocolAnalysisOutput'
 export * from './getProtocolModulesInfo'
 export * from './getLabwareOffsetLocation'
+export * from './getLabwareRenderInfo'
